commonize dialog components and sidebar query options (#314)
- Extract MoveCopyMediaDialog, SourceDeleteModal, AiTaggingModal to packages/ui
- Add projectsForMedia query option to packages/ui/query-options
- Update server-tauri-parity.md
