+++ .indexion/wiki/index.md
@@ -68,4 +68,3 @@ For specific topics, browse the **Categories** section.
 - pages add: packages-core
 - pages add: apps-xtracter
 - pages add: apps-cli
-
