+++ apps/server/src/components/media/media-sidebar/media-sidebar-content.tsx
@@ -1,5 +1,6 @@
 import type { MediaDetails } from "@solid-imager/core/domain/media/schemas";
 import { MediaSidebar as SharedMediaSidebar } from "@solid-imager/ui/media-sidebar";
+import { projectsQueryKeys } from "@solid-imager/ui/query-options";
 import { createQuery, useQueryClient } from "@tanstack/solid-query";
 import AiTaggingModal from "~/components/media/ai-tagging-modal";
 import {
@@ -19,9 +20,9 @@ import {
 	addProjectToMedia,
 	createProject,
 	fetchAllProjects,
-	fetchProjectsForMedia,
 	removeProjectFromMedia,
 } from "~/infrastructure/api-clients/projects-api";
+import { projectsForMediaQueryOptions } from "~/infrastructure/api-clients/queries/projects-query";
 
 type MediaSidebarProps = {
 	media: MediaDetails;
@@ -31,11 +32,9 @@ type MediaSidebarProps = {
 
 export default function MediaSidebar(props: MediaSidebarProps) {
 	const queryClient = useQueryClient();
-	const projects = createQuery(() => ({
-		queryKey: ["projectsForMedia", props.media.id],
-		queryFn: () =>
-			fetchProjectsForMedia(props.media.mediaSourceId, props.media.id),
-	}));
+	const projects = createQuery(() =>
+		projectsForMediaQueryOptions(props.media.mediaSourceId, props.media.id),
+	);
 	const allProjects = createQuery(() => ({
 		queryKey: ["allProjects"],
 		queryFn: fetchAllProjects,
@@ -51,7 +50,7 @@ export default function MediaSidebar(props: MediaSidebarProps) {
 
 	const invalidateProjectsForMedia = () =>
 		queryClient.invalidateQueries({
-			queryKey: ["projectsForMedia", props.media.id],
+			queryKey: projectsQueryKeys.forMedia(props.media.id),
 		});
 
 	return (
