+++ apps/server/src/components/media/move-copy-media-dialog.tsx
@@ -1,21 +1,6 @@
 import type { SafeMediaSource } from "@solid-imager/core/domain/sources/schemas";
-import { Button } from "@solid-imager/ui/button";
-import {
-	Dialog,
-	DialogContent,
-	DialogDescription,
-	DialogFooter,
-	DialogHeader,
-	DialogTitle,
-} from "@solid-imager/ui/dialog";
-import {
-	Select,
-	SelectContent,
-	SelectItem,
-	SelectTrigger,
-	SelectValue,
-} from "@solid-imager/ui/select";
-import { createResource, createSignal, Show } from "solid-js";
+import { MoveCopyMediaDialog as SharedMoveCopyMediaDialog } from "@solid-imager/ui/move-copy-media-dialog";
+import { createResource } from "solid-js";
 import { fetchMediaSources } from "~/infrastructure/api-clients/sources-api";
 
 type MoveCopyMediaDialogProps = {
@@ -27,8 +12,6 @@ type MoveCopyMediaDialogProps = {
 };
 
 export function MoveCopyMediaDialog(props: MoveCopyMediaDialogProps) {
-	const [targetSourceId, setTargetSourceId] = createSignal<string | null>(null);
-
 	const [sources] = createResource<SafeMediaSource[], boolean>(
 		() => props.open,
 		async (isOpen) => {
@@ -40,77 +23,21 @@ export function MoveCopyMediaDialog(props: MoveCopyMediaDialogProps) {
 		{ initialValue: [] },
 	);
 
-	const handleConfirm = () => {
-		const target = targetSourceId();
-		if (target) {
-			props.onConfirm(target);
-			props.onOpenChange(false);
-			setTargetSourceId(null);
-		}
-	};
-
-	const options = () =>
-		sources()
-			?.filter((s) => s.id !== props.currentSourceId)
-			.map((s) => ({ value: s.id, label: s.name })) ?? [];
+	const sourceOptions = () =>
+		(sources() ?? [])
+			.filter((s): s is SafeMediaSource & { id: string } => !!s.id)
+			.map((s) => ({ id: s.id, name: s.name }));
 
 	return (
-		<Dialog onOpenChange={props.onOpenChange} open={props.open}>
-			<DialogContent>
-				<DialogHeader>
-					<DialogTitle>
-						{props.mode === "copy" ? "Copy Media" : "Move Media"}
-					</DialogTitle>
-					<DialogDescription>
-						Select the destination source for this media item.
-						{props.mode === "move" &&
-							" The original item will be deleted after a successful copy."}
-					</DialogDescription>
-				</DialogHeader>
-
-				<div class="py-4">
-					<Show when={sources.loading}>
-						<p class="text-muted-foreground text-sm">Loading sources...</p>
-					</Show>
-					<Show when={sources.error}>
-						<p class="text-red-500 text-sm">Failed to load sources.</p>
-					</Show>
-					<Show when={!(sources.loading || sources.error)}>
-						<Select
-							itemComponent={(itemProps) => (
-								<SelectItem item={itemProps.item}>
-									{(itemProps.item.rawValue as { label: string }).label}
-								</SelectItem>
-							)}
-							onChange={(val) => setTargetSourceId(val?.value ?? null)}
-							options={options()}
-							optionTextValue="label"
-							optionValue="value"
-							value={
-								options().find((o) => o.value === targetSourceId()) ?? null
-							}
-						>
-							<SelectTrigger>
-								<SelectValue<{ label: string; value: string }>>
-									{(state) =>
-										state.selectedOption()?.label ?? "Select a source"
-									}
-								</SelectValue>
-							</SelectTrigger>
-							<SelectContent />
-						</Select>
-					</Show>
-				</div>
-
-				<DialogFooter>
-					<Button onClick={() => props.onOpenChange(false)} variant="outline">
-						Cancel
-					</Button>
-					<Button disabled={!targetSourceId()} onClick={handleConfirm}>
-						{props.mode === "copy" ? "Copy" : "Move"}
-					</Button>
-				</DialogFooter>
-			</DialogContent>
-		</Dialog>
+		<SharedMoveCopyMediaDialog
+			open={props.open}
+			onOpenChange={props.onOpenChange}
+			mode={props.mode}
+			currentSourceId={props.currentSourceId}
+			sources={sourceOptions()}
+			isLoading={sources.loading}
+			error={sources.error ? "Failed to load sources." : null}
+			onConfirm={props.onConfirm}
+		/>
 	);
 }
