+++ apps/server/src/infrastructure/api-clients/queries/projects-query.ts
@@ -1,5 +1,18 @@
-import { buildProjectsQueryOptions } from "@solid-imager/ui/query-options/projects-query";
-import { fetchAllProjects } from "../projects-api";
+import {
+	buildProjectsForMediaQueryOptions,
+	buildProjectsQueryOptions,
+} from "@solid-imager/ui/query-options/projects-query";
+import { fetchAllProjects, fetchProjectsForMedia } from "../projects-api";
 
 export const allProjectsQueryOptions = () =>
 	buildProjectsQueryOptions(fetchAllProjects);
+
+export const projectsForMediaQueryOptions = (
+	mediaSourceId: string,
+	mediaId: string,
+) =>
+	buildProjectsForMediaQueryOptions(
+		mediaSourceId,
+		mediaId,
+		fetchProjectsForMedia,
+	);
