+++ apps/tauri/src/components/media/move-copy-media-dialog.tsx
@@ -1,20 +1,6 @@
-import { Button } from "@solid-imager/ui/button";
-import {
-	Dialog,
-	DialogContent,
-	DialogDescription,
-	DialogFooter,
-	DialogHeader,
-	DialogTitle,
-} from "@solid-imager/ui/dialog";
-import {
-	Select,
-	SelectContent,
-	SelectItem,
-	SelectTrigger,
-	SelectValue,
-} from "@solid-imager/ui/select";
-import { createResource, createSignal, Show } from "solid-js";
+import type { SafeMediaSource } from "@solid-imager/core/domain/sources/schemas";
+import { MoveCopyMediaDialog as SharedMoveCopyMediaDialog } from "@solid-imager/ui/move-copy-media-dialog";
+import { createResource } from "solid-js";
 import { fetchMediaSources } from "~/infrastructure/api-clients/sources-api";
 
 type MoveCopyMediaDialogProps = {
@@ -26,8 +12,7 @@ type MoveCopyMediaDialogProps = {
 };
 
 export function MoveCopyMediaDialog(props: MoveCopyMediaDialogProps) {
-	const [targetSourceId, setTargetSourceId] = createSignal<string | null>(null);
-	const [sources] = createResource(
+	const [sources] = createResource<SafeMediaSource[], boolean>(
 		() => props.open,
 		async (isOpen) => {
 			if (!isOpen) {
@@ -37,63 +22,21 @@ export function MoveCopyMediaDialog(props: MoveCopyMediaDialogProps) {
 		},
 	);
 
-	const options = () =>
+	const sourceOptions = () =>
 		(sources() || [])
-			.filter((source) => source.id !== props.currentSourceId)
-			.map((source) => ({ value: source.id, label: source.name }));
-
-	const handleConfirm = () => {
-		const target = targetSourceId();
-		if (!target) {
-			return;
-		}
-		props.onConfirm(target);
-		props.onOpenChange(false);
-		setTargetSourceId(null);
-	};
+			.filter((s): s is SafeMediaSource & { id: string } => !!s.id)
+			.map((s) => ({ id: s.id, name: s.name }));
 
 	return (
-		<Dialog onOpenChange={props.onOpenChange} open={props.open}>
-			<DialogContent>
-				<DialogHeader>
-					<DialogTitle>{props.mode === "copy" ? "Copy Media" : "Move Media"}</DialogTitle>
-					<DialogDescription>Select the destination source for this media item.</DialogDescription>
-				</DialogHeader>
-				<div class="py-4">
-					<Show when={sources.loading}>
-						<p class="text-muted-foreground text-sm">Loading sources...</p>
-					</Show>
-					<Show when={!(sources.loading || sources.error)}>
-						<Select
-							itemComponent={(itemProps) => (
-								<SelectItem item={itemProps.item}>
-									{(itemProps.item.rawValue as { label: string }).label}
-								</SelectItem>
-							)}
-							onChange={(value) => setTargetSourceId(value?.value ?? null)}
-							options={options()}
-							optionTextValue="label"
-							optionValue="value"
-							value={options().find((option) => option.value === targetSourceId())}
-						>
-							<SelectTrigger>
-								<SelectValue<{ label: string; value: string }>>
-									{(state) => state.selectedOption()?.label ?? "Select a source"}
-								</SelectValue>
-							</SelectTrigger>
-							<SelectContent />
-						</Select>
-					</Show>
-				</div>
-				<DialogFooter>
-					<Button onClick={() => props.onOpenChange(false)} variant="outline">
-						Cancel
-					</Button>
-					<Button disabled={!targetSourceId()} onClick={handleConfirm}>
-						{props.mode === "copy" ? "Copy" : "Move"}
-					</Button>
-				</DialogFooter>
-			</DialogContent>
-		</Dialog>
+		<SharedMoveCopyMediaDialog
+			open={props.open}
+			onOpenChange={props.onOpenChange}
+			mode={props.mode}
+			currentSourceId={props.currentSourceId}
+			sources={sourceOptions()}
+			isLoading={sources.loading}
+			error={sources.error ? "Failed to load sources." : null}
+			onConfirm={props.onConfirm}
+		/>
 	);
 }
