+++ apps/tauri/src/infrastructure/api-clients/queries/projects-query.ts
@@ -1,4 +1,10 @@
-import { buildProjectsQueryOptions } from "@solid-imager/ui/query-options/projects-query";
-import { fetchAllProjects } from "../projects-api";
+import {
+	buildProjectsForMediaQueryOptions,
+	buildProjectsQueryOptions,
+} from "@solid-imager/ui/query-options/projects-query";
+import { fetchAllProjects, fetchProjectsForMedia } from "../projects-api";
 
 export const allProjectsQueryOptions = () => buildProjectsQueryOptions(fetchAllProjects);
+
+export const projectsForMediaQueryOptions = (mediaSourceId: string, mediaId: string) =>
+	buildProjectsForMediaQueryOptions(mediaSourceId, mediaId, fetchProjectsForMedia);
