+++ apps/tauri/src/routes/manager.tsx
@@ -13,11 +13,7 @@ import {
 	deleteCharacter,
 	updateCharacter,
 } from "~/infrastructure/api-clients/characters-api";
-import {
-	createIp,
-	deleteIp,
-	updateIp,
-} from "~/infrastructure/api-clients/ips-api";
+import { createIp, deleteIp, updateIp } from "~/infrastructure/api-clients/ips-api";
 import { orpc } from "~/infrastructure/api-clients/orpc-client";
 import {
 	createProject,
@@ -70,17 +66,8 @@ export default function ManagerPage() {
 	return (
 		<ManagerScreen
 			manager={manager}
-			renderMediaCard={(
-				media: Media,
-				selected: boolean,
-				onToggle: (mediaId: string) => void,
-			) => (
-				<MediaCardItem
-					media={media}
-					onToggle={onToggle}
-					selectable
-					selected={selected}
-				/>
+			renderMediaCard={(media: Media, selected: boolean, onToggle: (mediaId: string) => void) => (
+				<MediaCardItem media={media} onToggle={onToggle} selectable selected={selected} />
 			)}
 		/>
 	);
